# api/openapi-public.yaml
openapi: 3.0.3
info:
  title: Homeworldz Public API
  version: 0.1.0
  description: |
    The browser-facing API, consumed by the management site at
    `https://my.homeworldz.com` (see `web/src/config.js`) and by the Homeworldz
    client ([ADR 0030](../docs/adr/0030-client-architecture.md) and
    `docs/CLIENT2.md`). Served by the grid's `[website]` service, which listens
    on port 8003 by default.

    Distinct from `openapi.yaml`, which describes the internal grid API used by
    regions and tools rather than by browsers. The two share no credentials: a
    service token is not accepted here, and a website token is not accepted
    there.

    ## Conventions

    Every request body is one JSON object, at most 64 KiB, with **unknown
    fields rejected** — a misspelled property is `400 invalid_json` rather than
    a silently ignored field. Every error body is the `Error` schema, whose
    `code` is the stable machine-readable part; `message` is prose and may
    change.

    Two 404 codes mean different things and are deliberately distinguishable.
    `not_found` says the resource you named is absent or is not yours;
    `route_not_found` says this deployment serves no such endpoint, usually
    that it predates one. That way a client running ahead of a deployment can
    tell "not deployed" from "not allowed" without parsing prose.

    Requests may carry `X-Request-ID`; the value is echoed on the response, and
...

...
    `403 origin_forbidden`.

...
servers:
...
paths:
...
  /verifications/resend:
    post:
      operationId: resendVerification
      tags: [Registration]
      summary: Re-send a confirmation code
      description: |
        Mails a fresh confirmation code to the address on file.

        **The response is 202 whether or not the account exists**, whether or
        not it is already verified, and whether or not mail was sent, so this
        cannot be used to discover which accounts are registered.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userid]
              additionalProperties: false
              properties:
                userid:
                  type: string
                  description: The `first.last` login ID.
      responses:
        "202":
          description: Accepted. Returned for every outcome.
        "400":
          description: "`invalid_userid` (empty) or `invalid_json`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /password-resets:
    post:
      operationId: requestPasswordReset
      tags: [Authentication]
      summary: Request a password reset link
      description: |
        Sends a single-use reset link to the address on file for the named
        account. The link is only ever sent to that address, never to one
        supplied in the request.

        **The response is identical whether or not the account exists**, so this
        endpoint cannot be used to discover which accounts are registered. A
        caller learns only that the request was accepted.

        Rate limited. See [ADR 0034](../docs/adr/0034-password-recovery.md).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              additionalProperties: false
              properties:
                identifier:
                  type: string
                  minLength: 1
                  maxLength: 254
                  description: |
                    The login ID, in `first.last` form. This is the only accepted
                    identifier, and it is the same one login takes.

                    An email address is not accepted: nothing constrains it to a
                    single account, so it cannot name which password to change.
                    A display name is not accepted either — it can be changed,
                    while the login ID cannot. Case and spacing are normalized,
                    so `First Last` resolves to `first.last`.
      responses:
        "202":
          description: |
            Accepted. Returned whether or not an account matched, and whether or
            not mail was sent.
        "400":
          description: The request body was missing or malformed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429":
          description: Too many requests. The caller should retry later.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /password-resets/{token}:
    post:
      operationId: consumePasswordReset
      tags: [Authentication]
      summary: Set a new password using a reset token
      description: |
        Consumes a reset token and writes the new password. The token is
        single-use and expires one hour after it was issued.

        This performs exactly the same operation as an authenticated password
        change: it writes the password digests and nothing else. **Existing
        sessions are not ended** — a reset here means a forgotten password
        rather than a compromised account, and the reasoning is recorded in
        [ADR 0034](../docs/adr/0034-password-recovery.md).
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: The token from the emailed link.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              additionalProperties: false
              properties:
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
                  format: password
      responses:
        "204":
          description: The password was changed. The caller may now sign in with it.
        "400":
          description: |
            The token is unknown, already used, or expired, or the new password
            fails the length rules. **These are deliberately one response**:
            distinguishing them would say whether a token had ever existed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "429":
          description: Too many attempts. The caller should retry later.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
  /account:
    get:
      operationId: getAccount
      tags: [Account]
      summary: Read the signed-in account
      description: |
        Returns the caller's own identity, freshly loaded rather than decoded
        from the token, so a change made elsewhere is visible here immediately.
      security: [{ WebsiteToken: [] }]
      responses:
        "200":
          description: The caller's identity.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Identity" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "405": { $ref: "#/components/responses/MethodNotAllowed" }
        "503": { $ref: "#/components/responses/AccountStoreUnavailable" }
...
components:
...
  securitySchemes:
    WebsiteToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A website token from `POST /tokens` or `POST /verifications`. It is
        checked against the account's current authorization version on every
        request, so a password change, a privilege change, or a ban invalidates
        outstanding tokens at once. A region ticket from `POST /client/session`
        is not accepted here.
  responses:
...
    Unauthorized:
      description: |
        `unauthorized` — the token is absent, malformed, expired, or superseded
        by an authorization-version change. Carries `WWW-Authenticate: Bearer`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    MethodNotAllowed:
      description: "`method_not_allowed`. The `Allow` header lists what this route accepts."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: "`rate_limited`. Carries `Retry-After`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AccountStoreUnavailable:
      description: "`account_store_unavailable` — account storage is not configured or not reachable."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: A stable machine-readable slug. Branch on this, not on `message`.
        message:
          type: string
          description: Prose for a human. Wording may change without notice.
        field:
          type: string
          description: The offending request property, when one can be named.
...
    Privileges:
      type: string
      description: |
        A comma-separated list of privilege names, normalized to sorted and
        deduplicated on write. Empty means no privileges.

        `admin` expands to a curated common set — `users`, `bans`, `regions`,
        `map`, `deploy` — and deliberately **excludes** `undeploy`, which must
        be granted on its own. `super` grants everything, including itself.
      pattern: "^$|^[a-z][a-z0-9_-]*(,[a-z][a-z0-9_-]*)*$"
      example: users,bans
...
    Identity:
      type: object
      description: |
        An avatar identity. **Not a third-party profile**: `email` is on it, and
        no route serves an Identity to anyone but its owner or an administrator.
      required: [id, userid, displayName, rezDate, privs]
      properties:
        id: { type: string, format: uuid }
        userid:
          type: string
          description: The immutable `first.last` login ID.
        displayName: { type: string }
        email:
          type: string
          format: email
          description: Present for the account's own session and for administrators.
        rezDate:
          type: string
          format: date-time
          description: When the avatar was created.
        privs: { $ref: "#/components/schemas/Privileges" }
...